+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # v0.1.41 (2021-12-11)
 - **added:** `--upload-nfts` command that automates uploading of pairs of assets and metadata files, including updating metadata files with links to uploaded assets. You can now provide a glob pattern matching your asset file and the complete process will run, returning a manifest file with links to your uploaded metadata files that can be included in your on chain token metadata.
+- **added:** `--update-nft-status` command that reports on status for all nft uploads, assets, metadata and manifests
+for both assets and metadata.
 - **added:** `examples/upload_nfts`
 - **changed:** manifest paths now just include just file name instead of the full relative path
 - **changed:** sol service now has private rpc